Don't accept the scope as a parameter
The main goal of this library is to simplify the login flow as much as
possible. The returned data structure has a well defined, fixed set of
fields with the public components of a user's profile. There's no
point on being extra flexible with the scopes and then not using the
associated data.